% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFunctions.R
\name{current}
\alias{current}
\title{current}
\usage{
current(x)
}
\arguments{
\item{x}{Takes any value and outputs the current date.}
}
\value{
Outputs a string of the current date in the format Month Day-suffix Year
}
\description{
Outputs the current date in a suitable format for the journaling app.
}
