% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFunctions.R
\name{journal}
\alias{journal}
\title{journal}
\usage{
journal(journal)
}
\arguments{
\item{journal}{This argument opens up the journaling Shiny app.}
}
\value{
Returns a PDF file of the entered text of the journal prompt.
}
\description{
This functions opens up a Shiny app where users can journal inside a text box. Once the user is satisfied with their journal prompt, they can export it as a PDF file with the title as the current day they wrote the journal prompt.
}
