% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFunctions.R
\name{journal_prompt}
\alias{journal_prompt}
\title{journal_prompt}
\usage{
journal_prompt(text)
}
\arguments{
\item{text}{Takes any character vector to input as the journal prompt.}
}
\value{
A temporary PDF with all the text from the journal prompt in the correct format for the journaling app.
}
\description{
Calculates the number of pages necessary for the journal prompt, and creates a temporary PDF with the text correctly formatted.
}
