% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFunctions.R
\name{pomodoro}
\alias{pomodoro}
\title{pomodoro}
\usage{
pomodoro(pomodoro)
}
\arguments{
\item{pomodoro}{This argument opens up the pomodoro timer Shiny app.}
}
\value{
Starts, pauses and resets a pomodoro timer
}
\description{
This functions allows users to use a pomodoro timer for focused work or study. Users can specify differing work and break lengths.
}
