% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFunctions.R
\name{quotes_sample}
\alias{quotes_sample}
\title{quotes_sample}
\usage{
quotes_sample()
}
\arguments{
\item{x}{This argument takes any character vector, and will print one quote from the quotes vector.}
}
\value{
Returns the imputed character vector, in this case quotes.
}
\description{
Samples one quote at a time to be used in the Quotes app.
}
