% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFunctions.R
\name{required}
\alias{required}
\title{required}
\usage{
required()
}
\value{
Returns text in the console informing the user of what packages are being installed and loaded.
}
\description{
Installs and loads all necessary packages for the Shiny interface. Already installed packages are only loaded.
}
