% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFunctions.R
\name{shiny_app}
\alias{shiny_app}
\title{shiny_app}
\usage{
shiny_app(app)
}
\arguments{
\item{app}{This argument opens up the journaling Shiny app.}
}
\value{
Returns a user interface with tools.
}
\description{
This functions opens up a Shiny interface with interactive tools. Users can select tools to generate quotes, journal, use a pomodoro timer, and make a to do list. Once the user is satisfied with their journal prompt, they can export it as a PDF file.
}
