% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFunctions.R
\name{todolist}
\alias{todolist}
\title{To do list}
\usage{
todolist(todolist)
}
\arguments{
\item{todolist}{This argument opens up the to do list Shiny app.}
}
\value{
Opens an interactive to do list with checkboxes
}
\description{
This function allows users to imput specified tasks, which then populate to the right side of the screen. Users can then check off all done tasks which automatically move down to the bottom.
}
